lib: Implement converge_trees
Assume A, B, C are the divergent commits, P is the solution parents
(i.e. the parents chosen by converge_parents), and F is the evolution
fork point.

Notation: if N is an arbitrary commit then:
* N': commit N rebased on top of P
* N^: the merged parents of commit N.
* T(N): the MergedTree of commit N.

converge_trees returns the following merge:
T(F') + (T(A') - T(F')) + (T(B') - T(F')) + (T(C') - T(F'))

In words: it rebases F, A, B and C on top of the chosen parents (P) then
it merges the rebased trees of the divergent commits, using the tree of
the rebased evolution fork point as the base for this merge.

I left a TODO there: instead of using the evolution fork point as the
base, we should use the closest dominator value (see
find_dominator_value) as base. In most cases that should produce a
better merge. A later commit in this PR does that.

This commit also adds new basic APIs:
* Commit::parent_tree_no_resolve
* merge_commit_trees_no_resolve

Those new APIs are used to avoid resolving trees until the very end.